Move common components to separate directory
