+++ packages/frontend/src/components/common/BidListContainer.tsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const BidListContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  row-gap: 32px;
+  width: 100%;
+  margin: 0;
+  padding: 0;
+`
