+++ packages/frontend/src/components/common/BidsColumns.tsx
